Run all test in golem worker service
