--- golem-worker-service-base/config/worker-service.toml
@@ -1,40 +1,0 @@
-enable_tracing_console = false
-enable_json_log = false
-# workspace
-# environment
-port = 9000
-custom_request_port = 9001
-worker_grpc_port = 9092
-
-
-[redis]
-# host
-# port
-database = 0
-tracing = false
-pool_size = 8
-key_prefix = ""
-
-[redis.retries]
-max_attempts = 5
-min_delay = "100ms"
-max_delay = "2s"
-multiplier = 2
-
-[template_service]
-# host
-# port
-
-[template_service.retries]
-max_attempts = 3
-min_delay = "100ms"
-max_delay = "1s"
-multiplier = 3
-
-[worker_executor_client_cache]
-max_capacity = 1000
-time_to_idle = "4h"
-
-[routing_table]
-# host
-# port
