+++ golem-worker-service-base/src/service/mod.rs
@@ -1,4 +1,3 @@
 pub mod custom_request_definition_lookup;
 pub mod register_definition;
-
 pub mod error;
