+++ golem-worker-service/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod api;
+pub mod config;
 pub mod grpcapi;
 pub mod service;
-
 pub mod worker_request_to_http_response;
